requirements.txt
aiofile>=3.8.1
aiohttp>=3.8.0
aiohttp-socks>=0.8.0
colorama>=0.4.6
fake-useragent>=2.0.3
pycryptodome>=3.19
